+++ careMe/src/main/java/com/careme/controller/CasualBoardController.java
@@ -16,7 +16,7 @@ import com.careme.model.dto.BoardCommentDto;
 import com.careme.model.dto.PetSpeciesDto;
 import com.careme.model.dto.QuestionBoardDto;
 import com.careme.service.PetService;
-import com.careme.service.QuestionBoardService;
+import com.careme.service.QuestionBoardServiceImpl;
 import com.google.gson.Gson;
 
 
@@ -24,7 +24,7 @@ import com.google.gson.Gson;
 public class CasualBoardController {
 
 	@Autowired
-	QuestionBoardService bs;
+	QuestionBoardServiceImpl bs;
 	
 	@Autowired
 	QuestionBoardDao boardDao;
@@ -55,9 +55,10 @@ public class CasualBoardController {
 		bs.getCasualBoardViews(question_table_idx);
 		QuestionBoardDto mlist = bs.getCasualBoardContents(question_table_idx);
 		List<BoardCommentDto> clist = bs.getCasualBoardComments(question_table_idx);
-		
+		int commentCount = clist.size();
 		mav.addObject("mlist", mlist);
 		mav.addObject("clist", clist);
+		mav.addObject("commCount", commentCount);
 		mav.setViewName("casualBoardView/casualBoardContent");
 		return mav;
 	}
